Crate oak_io

Shared data structures and functionality for inter-node communication.

Modules

handle

Macros

handle_visit_blanket_impl

Structs

InitWrapper

A wrapper struct that holds an init message, plus the handle of a channel from which to read command messages. This is useful for patterns in which an Oak node needs some data at initialization time, but then handles commands of a different type and possibly coming from a different channel, to be processed in an event-loop pattern.

Message

A simple holder for bytes + handles, using internally owned buffers.

Receiver

Wrapper for a handle to the read half of a channel, allowing to receive data that can be decoded as bytes + handles via the Decodable trait.

Sender

Wrapper for a handle to the send half of a channel, allowing to send data that can be encoded as bytes + handles via the Encodable trait.

Enums

OakError

Generic Oak error.

OakStatus

Status values exchanged as i32 values across the Node Wasm interface.

Traits

Decodable

A trait for objects that can be decoded from bytes + handles.

Encodable

A trait for objects that can be encoded as bytes + handles.

Type Definitions

Handle

Handle used to identify read or write channel halves.