Module oak_io::handle

Structs

ReadHandle
Wrapper for a handle to the read half of a channel.

Receiver
Receiver handle for an Oak channel. This type is sent over the wire, but in types generated by prost it is replaced with oak::io::Receiver.

Sender
Sender handle for an Oak channel. This type is sent over the wire, but in types generated by prost it is replaced with oak::io::Sender.

WriteHandle
Wrapper for a handle to the send half of a channel.

Traits

HandleVisit
Visit all handles present in a type.

Functions

extract_handles
Return all handles in T.

inject_handles
Inject handles into a message.

Derive Macros

HandleVisit
Automatically derives the HandleVisit trait for structs and enums generated by prost.