Function oak_io::handle::inject_handles

pub fn inject_handles<T: HandleVisit>(
    msg: &mut T, 
    handles: &[Handle]
) -> Result<(), OakError>

Inject handles into a message.

If the number of handles provided is not exactly equal to the number of handles needed to fill the message, an error is returned.

Order is significant: handles are injected in field declaration order, recursing into nested structs before moving on to the next field.

use oak_io::{
    handle::{inject_handles, HandleVisit},
    Handle,
};

struct Thing {
    handle: Handle,
}

let handles = vec![42];
let mut thing = Thing { handle: 0 };

inject_handles(&mut thing, &handles).unwrap();

assert_eq!(thing, Thing { handle: 42 });